src/requirements.txt
aiosqlite==0.22.1
cartesia==3.0.2
dateparser==1.2.2
fastapi==0.136.1
google_api_python_client==2.195.0
httpx==0.28.1
ics==0.7.2
langchain_community==0.4.1
langchain_core==1.3.3
langchain_groq==1.1.2
langchain_openai==1.2.1
langchain_text_splitters==1.1.2
langgraph==1.1.10
protobuf==7.34.1
pydantic==2.13.3
python-dotenv==1.2.2
Requests==2.33.1
sarvamai==0.1.28
sounddevice==0.5.5
websockets==16.0
